fix register user bugs

+++ src/utils/cloudnary.js
@@ -2,9 +2,9 @@ import {v2 as cloudinary} from  "cloudinary"
 import fs from "fs"
 
     cloudinary.config({ 
-        cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-        api_key: process.env.CLOUDINARY_API_KEY,
-        api_secret:   process.env.CLOUDINARY_API_SECRET
+        cloud_name: "dgmy4ren9" ,
+        api_key: 473141645515214 ,
+        api_secret: "oDd2zkW85AWED6R225r9Oirk3tw"
     });
 
     const uploadOncloudinary = async (localFilePath) => {
@@ -16,11 +16,14 @@ import fs from "fs"
             })
 
             console.log("file is uploaded on cloudinary", response.url);
+            fs.unlinkSync(localFilePath)
             return response;
             
         } catch (error) {
+            if(fs.existsSync(localFilePath)){
             fs.unlinkSync(localFilePath) // remove the local saved 
             return null  
+            }
         }
     }
